The reapers have come
* All user-facing SEGS servers now have an assigned reaper that cull stale sessions
* In addition client connection state is propagated on change/connection stall up to Auth server
* Connection state has separate ids for MapServer and MapInstance, to allow multiple Instances on same MapServer

+++ Projects/CoX/Servers/AuthDatabase/AccountData.h
@@ -8,7 +8,7 @@ struct AuthAccountData
 {
     QString   m_login;
     uint8_t   m_access_level;
-    uint64_t  m_acc_server_acc_id=0;
+    uint32_t  m_acc_server_acc_id=0;
     QDateTime m_creation_date;
     bool valid() const { return m_acc_server_acc_id!=0; }
     bool isBlocked() const { return m_access_level == 0; }
